+++ src/styled-components.ts
@@ -1,24 +1,13 @@
-import * as React from 'react';
 import * as styledComponents from 'styled-components';
 
 import { ResolvedThemeInterface } from './theme';
 
 export { ResolvedThemeInterface };
 
-export type InterpolationFunction<P> = styledComponents.InterpolationFunction<P>;
-
-export type StyledFunction<T> = styledComponents.ThemedStyledFunction<T, ResolvedThemeInterface>;
-
-function withProps<T, U extends HTMLElement = HTMLElement>(
-  styledFunction: StyledFunction<React.HTMLProps<U>>,
-): StyledFunction<T & React.HTMLProps<U>> {
-  return styledFunction;
-}
-
 const {
   default: styled,
   css,
-  injectGlobal,
+  createGlobalStyle,
   keyframes,
   ThemeProvider,
 } = (styledComponents as any) as styledComponents.ThemedStyledComponentsModule<
@@ -54,7 +43,7 @@ export const media = {
   },
 };
 
-export { css, injectGlobal, keyframes, ThemeProvider, withProps };
+export { css, createGlobalStyle, keyframes, ThemeProvider };
 export default styled;
 
 export function extensionsHook(styledName: string) {